Move conductors into their own namespace
