+++ src/Streamnesia.Client/Hubs/OverlayHub.cs
@@ -1,5 +1,5 @@
 ﻿using Microsoft.AspNetCore.SignalR;
-using Streamnesia.Core;
+using Streamnesia.Core.Conductors;
 
 namespace Streamnesia.Client.Hubs;
