+++ src/Streamnesia.Client/Hubs/StatusHub.cs
@@ -1,6 +1,7 @@
 ﻿using Microsoft.AspNetCore.SignalR;
 using Streamnesia.Client.Models;
 using Streamnesia.Core;
+using Streamnesia.Core.Conductors;
 using Streamnesia.Core.Entities;
 
 namespace Streamnesia.Client.Hubs;
