+++ src/Streamnesia.Execution/ServiceCollectionExtensions.cs
@@ -2,6 +2,8 @@
 using System.Net.Sockets;
 using Microsoft.Extensions.DependencyInjection;
 using Streamnesia.Core;
+using Streamnesia.Core.Conductors;
+using Streamnesia.Execution.Conductors;
 
 namespace Streamnesia.Execution;
